Shift extractTaskNumber method from Class Task to Class Parser
