+++ src/main/java/omoh/Parser.java
@@ -1,4 +1,26 @@
 package omoh;
 
+import omoh.customexceptions.EmptyTaskNumberException;
+
 public class Parser {
+
+    //method that extracts the task number to mark or unmark
+    public static int extractTaskNumber(String input) throws NumberFormatException, EmptyTaskNumberException {
+        String keyword;
+        if (input.startsWith("mark")) {
+            keyword = "mark";
+        } else if (input.startsWith("unmark")) {
+            keyword = "unmark";
+        } else {
+            keyword = "delete";
+        }
+        String numberString = input.substring(keyword.length()).trim();
+        if (numberString.isEmpty()) {
+            throw new EmptyTaskNumberException();
+        }
+        int taskNumber = Integer.parseInt(numberString);
+        return taskNumber;
+    }
+
+
 }
