+++ src/main/java/omoh/tasktypes/Task.java
@@ -66,24 +66,6 @@ public class Task {
         fw.close();
     }
 
-    //method that extracts the task number to mark or unmark
-    public static int extractTaskNumber(String input) throws NumberFormatException, EmptyTaskNumberException {
-        String keyword;
-        if (input.startsWith("mark")) {
-            keyword = "mark";
-        } else if (input.startsWith("unmark")) {
-            keyword = "unmark";
-        } else {
-            keyword = "delete";
-        }
-        String numberString = input.substring(keyword.length()).trim();
-        if (numberString.isEmpty()) {
-            throw new EmptyTaskNumberException();
-        }
-        int taskNumber = Integer.parseInt(numberString);
-        return taskNumber;
-    }
-
     //method that modifies whether task is done or not done, depending on keyword mark or unmark detected
     public static void modifyDoneStateOrDelete(int taskNumber, String input) throws IndexOutOfBoundsException {
             if (input.startsWith("mark")) {
